# docs/swagger.yaml
definitions:
...
  model.CreateProjectRequest:
    description: CreateProjectRequest represents the inputs necessary to create a
      new project.
    properties:
      description:
        example: This is a new project
        format: string
        type: string
      id:
        example: 01979cde-6d91-7720-aaa5-d45582f94ac4
        format: uuid
        type: string
      name:
        example: New project name
        format: string
        type: string
    required:
    - description
    - name
    type: object
...
  model.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  model.ListProjectsResponse:
    description: ListProjectsResponse represents a list of users.
    properties:
      items:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      paginator:
        $ref: '#/definitions/model.Paginator'
    type: object
...
  model.Paginator:
    description: Paginator represents a paginator.
    properties:
      limit:
        example: 10
        format: int
        type: integer
      next_page:
        example: http://localhost:8080/users?next_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      next_token:
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      prev_page:
        example: http://localhost:8080/users?prev_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      prev_token:
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      size:
        example: 10
        format: int
        type: integer
    type: object
...
  model.Project:
    description: Project represents a project.
    properties:
      created_at:
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      description:
        example: This is a project
        format: string
        type: string
      disabled:
        example: false
        format: boolean
        type: boolean
      id:
        example: 01979cde-6d91-769f-8d3e-b04eeb538a83
        format: uuid
        type: string
      name:
        example: John
        format: string
        type: string
      system:
        example: false
        format: boolean
        type: boolean
      updated_at:
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  model.UpdateProjectRequest:
    description: UpdateProjectRequest represents the inputs necessary to update a
      project.
    properties:
      description:
        example: This is a new project data
        format: string
        type: string
      disabled:
        example: false
        format: boolean
        type: boolean
      name:
        example: New project name
        format: string
        type: string
    type: object
...
info:
  contact:
    email: info@qu3ry.me
    name: API Support
    url: https://qu3ry.me
  description: |-
    This is a service template for building RESTful APIs in Go.
    It uses a PostgreSQL database to store user information.
    The service provides:
    - CRUD operations for users.
    - Health and version endpoints.
    - Configuration using environment variables or command line arguments.
    - Debug mode to enable debug logging.
    - TLS enabled to secure the communication.
  title: Go REST API Service Template
  version: v1
paths:
...
  /projects:
    get:
      description: Retrieve paginated list of all projects in the system
      operationId: 019797e6-138a-7cf1-b7bb-fa9c5e168c49
      parameters:
      - description: 'Comma-separated list of fields to sort by. Example: first_name
          ASC, created_at DESC'
        format: string
        in: query
        name: sort
        type: string
      - description: 'Filter field. Example: id=1 AND first_name=''John'''
        format: string
        in: query
        name: filter
        type: string
      - description: 'Fields to return. Example: id,first_name,last_name'
        format: string
        in: query
        name: fields
        type: string
      - description: Next cursor
        format: string
        in: query
        name: next_token
        type: string
      - description: Previous cursor
        format: string
        in: query
        name: prev_token
        type: string
      - description: Limit
        format: int
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListProjectsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project with specified configuration
      operationId: 019797e6-138a-7d00-98db-740f21794f11
      parameters:
      - description: Create Project Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Create project
      tags:
      - Projects
  /projects/{project_id}:
    delete:
      description: Remove a project permanently from the system
      operationId: 019797e6-138a-7cf4-8694-e4611baded39
      parameters:
      - description: The project id in UUID format
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Delete project
      tags:
      - Projects
    get:
      description: Retrieve a specific project by its unique identifier
      operationId: 019797e6-138a-7d04-8db3-1d4755b25db3
      parameters:
      - description: The project id in UUID format
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Get project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Modify an existing project by its ID
      operationId: 019797e6-138a-7cf8-9887-e4c44ad0ae19
      parameters:
      - description: The project id in UUID format
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Update Project Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Update project
      tags:
      - Projects
...
swagger: "2.0"